+++ nft-v2/test/.eslintignore
