+++ nft-v2/test/fileExtReplacer.js
